Re-updated the way we load script-files for the browser.
Loading of descriptions in the browser with jQuery was causing a CORS-exceptions, which really deemed the browser option as useless. reverted back to loading with requirejs in the browser, but sticking to jQuery in the runner.
There might be more elegant ways of doing this, but it works and can be improved if needed :) closes #66