Fixing so we don't get duplicate suites - taking out a check for running things twice.
This needs looking into, but effectively - VS2013 has a temporary filename for any file being edited while editing it and then renames it / copies it and replaces the file you're supposed to be editing.  Weird stuff! This causes a rename and then a change!  Also worth mentioning that I've turned off raising of events while a run is running, which should limit things to one run!